♻️ separate form as component, split children input
