+++ src/Pages/Adopt/index.js
@@ -1,0 +1,1 @@
+export { Adopt } from "./Adopt";
